+++ config/rollup.config.js
@@ -1,9 +1,9 @@
+import babel from '@rollup/plugin-babel';
+import commonJS from '@rollup/plugin-commonjs';
+import { nodeResolve } from '@rollup/plugin-node-resolve';
 import bundlesize from 'rollup-plugin-bundle-size';
-import resolve from 'rollup-plugin-node-resolve';
-import commonJS from 'rollup-plugin-commonjs';
-import babel from 'rollup-plugin-babel';
-import { uglify } from 'rollup-plugin-uglify';
-import pkg from './package.json';
+import { terser } from 'rollup-plugin-terser';
+import pkg from '../package.json';
 
 export default [
   // Browser-friendly UMD build
@@ -15,14 +15,18 @@ export default [
       format: 'umd',
     },
     plugins: [
-      resolve(),
-      commonJS({ include: 'node_modules/**' }),
-      babel({ exclude: 'node_modules/**' }),
-      uglify({ sourcemap: false }),
+      nodeResolve(),
+      commonJS({
+        include: 'node_modules/**',
+      }),
+      babel({
+        babelHelpers: 'bundled',
+        exclude: 'node_modules/**',
+      }),
+      terser(),
       bundlesize(),
     ],
   },
-
   // CommonJS (for Node) and ES module (for bundlers) build.
   // (We could have three entries in the configuration array
   // instead of two, but it's quicker to generate multiple
@@ -45,7 +49,10 @@ export default [
       },
     ],
     plugins: [
-      babel({ exclude: 'node_modules/**' }),
+      babel({
+        babelHelpers: 'bundled',
+        exclude: 'node_modules/**',
+      }),
       bundlesize(),
     ],
   },