+++ release.config.js
@@ -1,31 +1,32 @@
-module.exports = {
+const mainConfig = {
   branches: [
     'master',
-    {
-      name: 'develop',
-      prerelease: 'beta',
-    },
-    {
-      name: 'alpha',
-      prerelease: 'alpha',
-    },
+    { name: 'develop', channel: 'beta', prerelease: 'beta' },
   ],
   repositoryUrl: 'https://github.com/nugit/datetime-utils.git',
   plugins: [
-    '@semantic-release/commit-analyzer',
-    '@semantic-release/release-notes-generator',
-    '@semantic-release/changelog',
     [
-      '@semantic-release/git',
+      '@semantic-release/commit-analyzer',
       {
-        assets: [
-          'CHANGELOG.md',
-        ],
-        // eslint-disable-next-line no-template-curly-in-string
-        message: 'chore(release): ${nextRelease.version}\n\n${nextRelease.notes}',
+        preset: 'conventionalcommits',
+      },
+    ],
+    '@semantic-release/release-notes-generator',
+    [
+      '@semantic-release/changelog',
+      {
+        changelogFile: 'CHANGELOG.md',
       },
     ],
     '@semantic-release/npm',
+    [
+      '@semantic-release/git',
+      {
+        assets: ['CHANGELOG.md', 'package.json'],
+        // eslint-disable-next-line no-template-curly-in-string
+        message: 'chore(release): ${nextRelease.version}',
+      },
+    ],
     [
       '@semantic-release/github', {
         assets: [
@@ -45,3 +46,33 @@ module.exports = {
     ],
   ],
 };
+
+const { execSync } = require('child_process');
+const { createHash } = require('crypto');
+
+const branch = execSync('git branch --show-current').toString().trimEnd('\n');
+const channel = `alpha.${createHash('md5').update(branch).digest('hex')}`;
+
+const localConfig = {
+  branches: [
+    'master',
+    { name: 'develop', channel: 'beta', prerelease: 'beta' },
+    {
+      name: branch,
+      channel,
+      prerelease: channel,
+    },
+  ],
+  repositoryUrl: 'https://github.com/nugit/datetime-utils.git',
+  plugins: [
+    [
+      '@semantic-release/commit-analyzer',
+      {
+        preset: 'conventionalcommits',
+      },
+    ],
+    '@semantic-release/npm',
+  ],
+};
+
+module.exports = process.env.LOCAL_RELEASE ? localConfig : mainConfig;